virtme: propagate /proc/sys/fs/nr_open from host to the guest
Some distro, such as Fedora or CachyOS, are using a higher limit of max
open files, beyond the kernel default. Consequently, some applications
may expect to operate under this increased limit and, as a result, we
may hit EPERM errors in the virtme-ng guest.

Example:

 $ sudo su -
 sudo: pam_open_session: Permission denied
 sudo: policy plugin failed session initialization

 # strace -f -e prlimit64,setrlimit sudo /bin/true
 ...
 prlimit64(0, RLIMIT_NOFILE, {rlim_cur=1024, rlim_max=2048*1024}, NULL) = -1 EPERM (Operation not permitted)
 ...
 # cat /proc/sys/fs/nr_open
 1048576

Fix this by propagating the value from /proc/sys/fs/nr_open from the
host to the vng guest.

This fixes one of the issues reported in #75.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>