init: allow to mount legacy cgroupfs (v1)
Allow to mount legacy cgroup when SYSTEMD_CGROUP_ENABLE_LEGACY_FORCE=1
is passed to the kernel boot command options.

This mimics systemd's behavior, as specified in the systemd
documentation:
```
CHANGES WITH 256 in spe:
...
    * Support for cgroup v1 ('legacy' and 'hybrid' hierarchies) is now
      considered obsolete and systemd by default will refuse to boot under it.
      To forcibly reenable cgroup v1 support, SYSTEMD_CGROUP_ENABLE_LEGACY_FORCE=1
      must be set on kernel command line.
...
```

Moreover, with cgroup v1, only try to mount the same legacy cgroup
subsystems supported by systemd (also from the systemd doc):
```
* on cgroup v1: `cpu`, `cpuacct`, `blkio`, `memory`, `devices`, `pids`
```

This was discussed in #84.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>