config: always append virtme configs
Otherwise, the 'virtme-configkernel --update' command doesn't really
make sense if it doesn't update an existing kernel config file.

Note that 'vng -b' and 'vng -k' will call 'virtme-configkernel --update'.
It is not clear if the kconfig update is skipped on purpose or not, but
it looks confusing not to do anything.

Fixes: 6ade5a7 ("virtme: improve configkernel")
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>