config: option not to update if there is a config file
The '--update' option of virtme-configkernel is there to "update
existing config for virtme" according to the help message.

vng was calling 'virtme-configkernel --update', but the goal was not to
modify the config file if it was present, just in case it has been
modified in between two build calls. To support that, the '--update'
option had been modified ... not to update the config file if it was
present, in contradiction with the option name and its 'help' text.

A new '--no-update' parameter is now available: if the config file
exists, nothing is done. 'vng' is then now calling:

  virtme-configkernel --defconfig --no-update

So if the config file exists, the config is not modified, but a message
is now printed in stdout (only visible when vng is used with --verbose).

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>