qemu: remove unnecessary '-serial none' arg
When not using the --graphics mode, or without the script options, QEmu
was always launched with:

  -serial none -serial chardev:console

'-serial none' means "don't allocate this serial device". Here, it is
directly overridden by another serial. Best not to declare '-serial none'
then.

Since QEmu 8.2.2, having '-serial none -serial <something>' is no longer
supported, and it is then recommended to remove the unnecessary '-serial
none' [1]. It continues to work on previous versions (tested on 8.2.1).

Reported-by: @tehcaster
Closes: #97
Link: https://lore.kernel.org/qemu-devel/20240122163607.459769-2-peter.maydell@linaro.org/ [1]
Suggested-by: @ishitatsuyuki
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>