virtme-ng: provide a man page for vng
Automatically generate a manpage via argparse-manpage.

We can improve the content of the man page later in the future, but for
now it has all the basic information about the command line options and
their description, that can be useful when it's not possible to access
the main README.md on the github page.

This closes issue #89.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>