virtme-ng: check validity of rwdir paths
When using the rwdir option in the format --rwdir=/guestpath=/hostpath,
it is crucial to ensure that both guestpath and hostpath are valid
before starting the instance. Instead of crashing, the system should
report a proper error message if the paths are not valid.

To achieve this, verify that the specified host path is accessible and
check whether the guest path exists or can be created within a valid
overlay.

This addresses issue #117.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>