virtme: introduce option --gdb
Introduce an option to easily attach a gdb session to a running
virtme-ng instance started with --debug.

With this option vng will simply act as a wrapper to gdb, with some
extra logic to determine if a vmlinux is available (either locally built
or in the standard distro location if a pre-compiled kernel is
executed).

This option can be used together with --run / -r to help vng recognize
the proper kernel version and potentially find the proper kernel
symbols.

Example:

 - Start a vng instance in debug mode, using the system's kernel
   6.8.0-31-generic:

   $ vng -r 6.8.0-31-generic --debug

 - On another terminal start the gdb session:

   $ vng -r 6.8.0-31-generic  --gdb
   ...
   0xffffffff82222e5b in pv_native_safe_halt () at /build/linux-yrLejD/linux-6.8.0/arch/x86/kernel/paravirt.c:128
   (gdb)

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>