virtme-configkernel: 'O=' takes precedence over KBUILD_OUTPUT
For the moment, it is required to set KBUILD_OUTPUT env var, and not
just passing O=, because only KBUILD_OUTPUT env var is used to find the
path to the .config file.

If O= is set in the env vars for make, then use it to override
KBUILD_OUTPUT env var, as it is done on the kernel side.

Reported-by: @jimc
Closes: #127
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>