vng-run: get kdir from O=outdir
If O=outdir is given as an arg to vng -b, it will build the whole
kernel in outdir.  Then a following vng O=outdir will boot, but
cannot find the modules, and lsmod shows the empty set.

So for ``vng O=outdir -r``, adjust the qemu --kdir arg from the
default: ./ to ./outdir.

TBD: this is currently not done for -r <value>, since the value may be
a version-tag to be downloaded.  This would need separate
consideration.

NB: outdir is path relative, and would break if given an absolute
path.  This likely means that O=jimc@remote:/data/kerneldump/foo would
break badly.  I don't forsee a practical usability problem with this
limitation, but Im not that kind of user.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>