version: generate verison string based on git information
If virtme-ng is installed from the git repository add git information to
the version string.

If, for any reason, we fail to determine the git information (e.g., when
building virtme-ng from a distro source package), fallback to the main
packaging version, typically updated manually every time we cut a new
release.

Suggested-by: Jim Cromie <jim.cromie@gmail.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>