version: try to get version from pip
If we can't determine any version info from `git describe` try to get
the version from pip, otherwise fallback to the hard-coded package
version.

Also force to use the git version string both when we are installing
(setup.py) the package or when we are directly running vng from the
source repo.

Signed-off-by: Andrea Righi <righi.andrea@gmail.com>