version: fallback to PKG_VERSION if git describe failed
Now that 'git describe' is only used for local version, it means that in
case of issue with the git command, the version from pip cannot be
picked as an alternative, because it might be different from the local
one being used here. A fallback to the hard-coded package version seems
to be the only alternative if 'git describe' cannot be used.

This cover the case of someone using the source code from a tarball,
without git support, but potentially having a different version already
installed with pip.

Fixes: fb9f6be ("version: try to get version from pip")
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>