virtme: get rid of virtme/scripts symlink
There is no reason to maintain a symlink from virtme/scripts -> bin, we
can just run the scripts directly from bin, when vng is executed
directly from the source dir (otherwise all the scripts will be
available in the PATH and we don't need to do anything special).

This can solve potential issues with packaging that may have problems
to preserve the symlinks.

Signed-off-by: Andrea Righi <righi.andrea@gmail.com>