Don't require ".git" to be a directory
Running "vng --build --build-host user@host" within my git checkout, I
get:

    error: --build-host can be used only on a kernel git repository

This is because my checkout is actually a "worktree". Git worktrees
allow multiple independent checkouts of the same git repository to exist
on the filesystem, sharing the Git metadata. The ".git" for these
repositories is a file containing the path to the "real" directory.
Relax the detection condition so a file named ".git" is allowed.

Signed-off-by: Stephen Brennan <stephen@brennan.io>