virtme: Enable -r argument to work on more architectures
Using vng -r argument was not working properly on ppc64le and s390x,
since the code to find the currently running kernel was searching for
vmlinuz.

For now, create a new entry on Arch class to specify the name of the
image to look to, with the default still being vmlinuz. To make ppc64le
to work, add vmlinux as ppc64le for now.

vng -r starts to work on ppc64le after this change.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>