virtme-ng: handle undefined release with --root
When the --root option is used to automatically create a new chroot from
scratch, virtme-ng attempts to populate the chroot using the latest
Ubuntu cloud image.

The specific Ubuntu release is determined by running lsb_release on the
host system, aiming to match the host environment as closely as
possible. However, if the host is not running Ubuntu, this command may
either fail or return "n/a" (e.g., on Arch Linux).

In such cases, treat the "n/a" result as an error and prompt the user to
manually specify the target Ubuntu release for initializing the chroot.

Fixes: 97dd475 ("Add hint when trying to create root filesystem outside a Debian host")
Signed-off-by: Andrea Righi <andrea.righi@linux.dev>