Revert "Modernize build system"
Adjusting the build system to a more modern way is a good plan, but this
change ended up introducing a few regressions in the packaging: bash
completion and default config file are not shipped anymore in distro,
causing some issues.

Therefore, let's revert this change for now and try to re-introduce
these changes in a more incremental way, making sure that we can keep
shipping all the expected files.

This has been discussed in issue #181.

Signed-off-by: Andrea Righi <andrea.righi@linux.dev>