net: support multiple interfaces
virtme-run supports multiple --net arguments to setup virtio-net
devices, but the init program was only starting the DHCP client for one
of them: the first one by alphabetical order.

Now, the DHCP client is started for each interface, so each of them can
have an assigned IP address.

All these DHCP clients are started in parallel, and the script wait for
them to finish to continue, instead of doing them one by one.

This also includes the modification in the new init.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>