vsock: mount virtme_vsockmount if needed
This will be used to get access to a script that will setup the remote
shell: HOME dir, log with the right user, set tty, etc.

This script will need to be created when a new connection is needed, to
get access to the terminal dimension, and maybe more. Because of that,
it is needed to have a way to share such info from the host to the VM.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>