vsock: connect: generate helper script
Directly connecting to the remote shell using socat launching 'bash'
remotely doesn't give a nice user experience: env vars like $HOME are
not propagated, the terminal size is set to 80x24, the command that is
typed is not visible, the directory is not the one picked with --cwd,
etc.

To give a better user experience, the '--vsock-connect' option is now
creating a script that is setting up the missing pieces. This script is
executed when the VM is launched with '--vsock' without any arguments.

Note that it is required to pass info from the host, when launching
'--vsock-connect', to the VM, e.g. for the terminal size. This is done
via the script that is executed once connected to the vsock inside the
VM. The VM needs access to this script, and the host needs to be able to
modify it. Such script is then created in /tmp/virtme-vsock/<CID>.sh on
both the host and the VM. If the whole filesystem is mounted (default),
this file will be accessible from the VM. If not, a mount point will be
added to access the parent directory from the VM.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>