make: build virtme-ng-init by default
As mentioned in the README file, the 'make' command should only build
virtme-ng-init. Building is what 'make' usually does, and not installing
stuff.

Since commit 85a9b42 ("Make it build on macOS"), 'make' was building
virtme-ng-init, but it was also trying to install it with other files
elsewhere on the system, what 'make install' does. That's because make
was using the first target as the default one, so the 'install' one.

Instead of directly calling 'cargo' like it was the case before, the
'build' part of setup.py is called. This will generate files in the
build dir that are probably not needed, but it is quick, and it will
also build virtme-ng-init properly, what we want.

While at it:

- Mention all targets that are not linked to files in .PHONY.

- Only generate the init file when needed: when the .rs files have been
  modified.

- Initialise the submodule if the .rs files are not present.

- A new 'clean' target can help to force rebuilding (or just to clean,
  even if it looks like it missed many generated files...).

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>