vsock: connect: support commands without tty
It might be interesting to run 'vng --vsock-connect <cmd>' in
environments without any TTY support, e.g. what automated scripts could
do in a CI environment.

In this case, it is not possible to use /dev/pts/X for the input file
for socat. Instead, stdin (-) can be used as a fallback.

While at it, no need to execute the 'stty' command in this case,
especially because the 'echo' parameter will duplicate the command that
is typed via stdin.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>