vsock: rename options to --client/--server
The idea is to have other types of remote connections later on, e.g.
SSH.

So to support future types later on, the syntax is now:

  vng --server|--client [console|ssh] [--port PORT] [--remote-cmd COMMAND]

The 'ssh' option is not available yet, but by default, 'console' will be
used, so users can just do:

  $ vng --server
  $ vng --client  # in another terminal

A new group for the parser has been added to avoid confusions: --port
and --remote-cmd are linked to --client/--server.

While at it, the code for the console client/server has been moved to
dedicated functions, grouped together. It will also be easy to support a
new type later on, simply by calling a different function.

Note that this should only be a refactoring, the behaviour, apart from
different option name, should not be modified.

The variables names have been renamed too. 'vsock' is now only used in
virtme(-ng)-init and in the socat command.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>