virtme: enable ssh support
Add support for specifying `--server ssh`, to automatically start and
configure sshd in the virtme-ng guest, and `--client ssh` to connect to
a virtme-ng guest via SSH.

The server's port can be customized with the `--port NUM` option
(default is 2222).

Example usage (start a vng instance with sshd enabled):

 $ vng --server ssh --port 2222

In another shell session on the host, connect to the running instance as
a client:

 $ vng --client ssh --port 2222

Note that it's also possible to use ssh or scp directly, the --client
option is provided for convenience and to be consistent with the vsock
option.

Signed-off-by: Andrea Righi <arighi@nvidia.com>