sshd: look for overlayfs before modifying auth keys
Only looking if the rootfs is in read-only might not be enough, e.g. if
the home directory is mounted in RW.

Instead, we can simply check in which file-system the authorized_keys
file is in: if it is in an overlayfs, it is safe.

Note that the '%T' format could be used with stat, but it looks like it
is not working with busybox.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>