vng: run: support O= in interactive mode
When vng was used with O=my_dir/ to launch a VM, the VM stopped as soon
as the boot was done without any message, e.g.

  host: $ vng O=.virtme/build
  host: $

That's because O=my_dir/ was used as a command to execute in a
non-interactive mode: so simply setting an env var, and exiting.

Now, when O=... is passed, it is used to set the build dir where the
kernel image is, and this command is removed from the exec list. In
other words, this command will start a VM and start a prompt:

  host: $ vng O=.virtme/build
    vm: $

Fixes: afbec37 ("vng-run: get kdir from O=outdir")
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>