vng: support --debug with --qemu-opts
When running with both --debug and --qemu-opts flags, one space
character was missing in the --qemu-opts passed on the command line to
virtme-run.

This could be fixed by adding a space as a prefix for the QEmu arguments
added for the --debug case, but it sounds better to add the --qemu-opts
flags at the end, so the debug args don't take precedence.

Closes: #230
Reported-by: @manufactory
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>