vng: support --nvgpu and --qemu-opts
When passing arguments to virtme-run, --qemu-opts has to be the last
one, because it will take all args set after that.

In other words, --nvgpu was not supposed to be passed after --qemu-opts.

While at it, re-order functions and add a comment to avoid re-doing the
same mistake.

Note that an alternative would be to call virtme-run with --qemu-opt
instead of --qemu-opts (with 's'). But a comment should be enough.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>