config: create build dir if defined
Before this patch, if KBUILD_OUTPUT or O env vars were set, but the
mentioned directory didn't exist, the script was first running ...

  $ KBUILD_OUTPUT=(...) make defconfig

... generating the .config in the build dir, but appending the extra
kconfigs in the source dir.

The restriction for the build dir doesn't seem to be needed. Instead,
the directory is created if it doesn't already exist. If it cannot be
created, an error is returned.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>