init: clean up previous X11 state
When running vng in graphic mode (-g), recent versions of Xorg may fail
to start with the following error:

 _XSERVTransSocketUNIXCreateListener: ...SocketCreateListener() failed
 _XSERVTransMakeAllCOTSServerListeners: server already running
 (EE)
 Fatal server error:
 (EE) Cannot establish any listening sockets - Make sure an X server isn't already running(EE)

This can happen because the host's /tmp is exposed to the guest and Xorg
might get confused finding a previous state.

Fix by removing the Xorg state files before starting the X server in the
vng guest.

Fixes: da5b586 ("virtme-ng: export real /tmp to guest")
Signed-off-by: Andrea Righi <arighi@nvidia.com>