setup.py: Build a static virtme-ng-init
When using vng on a filesystem different than the host's filesystem
(when `--root DIR` is used) we may have a different glibc version or
different libraries in general, that are not compatible with the
pre-built virtme-ng-init.

To fix this build a static virtme-ng-init, so that it can be used across
any rootfs.

This fixes #243.

Signed-off-by: Andrea Righi <arighi@nvidia.com>