systemd: Add support for separate chroot
Add a new `--systemd` optional argument which, when used in addition to
`--root` pointing to an external filesystem (i.e. not "/"), will execute
`/sbin/init` with a customized "serial-getty@.service" which in turn calls
the usual "virtme-init" script. This is the default service that systemd
targets when the kernel has "console=ttyS0".

The "virtme-init" script receives its environment variables from a
"virtme-serial-getty.env" file generated at runtime, which is made
available to systemd by setting $SYSTEMD_UNIT_PATH pointing to the
guesttools mount path. That is where the environment file will be, along
with the custom "serial-getty@.service".

Systemd will setup the system as usual, therefore we must bypass many
things that "virtme-init" usually needs to do when `--systemd` is not used.
For this, add "virtme_systemd=1" to the environment.

Signed-off-by: Ricardo B. Marlière <ricardo@marliere.net>