virtme-ng: Do not truncate stdout/stderr on process exit
When an external command is executed by vng, its stdout or stderr might
get truncated if the child process exits before all the output is
consumed.

This can happen because data may still be buffered in the stdout or
stderr pipes after the child terminates.

Fix this by making sure both stdout and stderr are fully drained, even
after the process has exited, ensuring that all output is captured and
correctly reported.

Example:

 - before this change:

   $ vng -vb --config /no-file
   cmd: virtme-configkernel --defconfig --verbose --no-update --custom /no-file
   usage: virtme-configkernel [-h] [--arch ARCHITECTURE]
                              [--cross-compile CROSS_COMPILE_PREFIX]

 - after this change:

   $ vng -vb --config /no-file
   cmd: virtme-configkernel --defconfig --verbose --no-update --custom /no-file
   usage: virtme-configkernel [-h] [--arch ARCHITECTURE]
                              [--cross-compile CROSS_COMPILE_PREFIX]
                              [--custom CUSTOM] [--configitem CONFITEM]
                              [--no-update] [--verbose] [--allnoconfig |
                              --defconfig | --update]
                              [envs ...]
   virtme-configkernel: error: argument --custom: '/no-file' is not a valid file.

Signed-off-by: Andrea Righi <arighi@nvidia.com>