vng: build/quote: quotes envs and args
Just in case there are spaces, e.g.

  vng -v CC="ccache gcc"

All arguments coming from the command line are now quoted. Use an array
instead of a long string that is (potentially not correctly) split
later.

While at it, move MAKE_COMMAND to the only place where it is used. If an
array is used, probably better like than than having to deal with
copies.

Closes: #286
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>