README: example with KBUILD_OUTPUT
KBUILD_OUTPUT and O= can be used.

It might be clearer to deal with the env vars than the special O=.

While at it, put O= at the end, after --, so other commands can be used.
Also, in my test, it is no longer needed to create the build dir before.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>