systemd: Enable user to customize masks
Due to the differences in root filesystems and distros, the user might need
to mask away some other services to have a functioning system using systemd
as init. Enable the user to do so through the configuration file, instead
of needing to rely on using `--append "systemd.mask={}"` in the command
line.

Signed-off-by: Ricardo B. Marlière <rbm@suse.com>