github: rust: use one pre-commit command
The goal of these steps is to check everything related to the rust files
is OK. Instead of listing each cargo-related operation from the
pre-commit config file, the "--files" argument can be used to obtain the
same goal, but in one operation.

This should simplify the maintenance when the pre-commit file is
modified, not to have to keep both synced.

Note that the action will also run extra generic hooks (trailing
whitespace, etc.), but I guess that's fine. Also, the extra repos are
cached, so that seems to be a good enough trade-off.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>