commands: run.py: Resolve symlinks on get_kernel_version
In distributions like openSUSE/SLE, vmlinuz files can be just symlinks
pointing to the real file elsewhere, like the example below:

$ file /boot/vmlinuz
/boot/vmlinuz: symbolic link to vmlinuz-6.16.3-1-default

In such cases, the detection of the kernel version may fail to find
the kernel version, and thus not setting up the kernel modules, making
the boot to fail. This happens because one of the detections rely on
the output of the file command, which shows the kernel version if the
file is a regular file:

$ file /usr/lib/modules/6.16.3-1-default/vmlinuz | grep version
/usr/lib/modules/6.16.3-1-default/vmlinuz: Linux kernel x86 boot executable,
bzImage, version 6.16.3-1-default...

Now, if the user runs vng -r /boot/vmlinuz, it can fail because the output
of the file command is different:

$ file /boot/vmlinuz
/boot/vmlinuz: symbolic link to vmlinuz-6.16.3-1-default

The fix resolves the symlink, so vng can have the output it wants. It
doesn't change the rationale of the detect, just avoids one corner case.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>