virtiofsd: run with -o cache=auto in --rw mode
When --rw is used, the user expects a FS synchronisation between the
guest and the host. In this case, virtiofsd should not be launched with
"-o cache=always".

When starting vng without --rw, it seems OK to consider that the guest
has a "fork" version of the host's filesystem: files modified in the
guest are not reflected on the host (except if --rwdir is used). Then,
it is OK to say that files read in the guest keep their old version when
they are updated on the host. When --rw is used, it is different: this
synchronisation is expected.

Now, when vng is launched with "--rw", virtiofsd is executed with
cache=auto.

Note: while at it, re-order virtiofsd arguments: {stderr} is kept at the
end, just to avoid '<cmd> <param> 2>/dev/null <param>'. Even if that's
OK, that looks strange.

Closes: #353
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>