virtme-init: Improve nested virtme-ng performance
Nested virtme-ng instances are very slow out of the box:

    host:linux$ ../virtme-ng/vng --build
    host:linux$ ../virtme-ng/vng

    [    0.019238] Booting paravirtualized kernel on KVM
    [    0.253527]     virtme_user=iii

    virtme-ng:linux$ ../virtme-ng/vng

    [    0.236793] Booting paravirtualized kernel on bare hardware
    [    7.396265]     virtme_user=iii

This is because they use TCG instead of KVM. This is, in turn, because
/dev/kvm is not accessible due to having root:root ownership instead of
root:kvm.

Setting the ownership is implemented in
/lib/udev/rules.d/50-udev-default.rules as follows:

    KERNEL=="kvm", GROUP="kvm", MODE="0660", OPTIONS+="static_node=kvm"

virtme-ng starts udev with the --resolve-names=never option, which
makes it ignore the GROUP= assignment [1].

virtme-ng uses this option from the beginning, but it's not really
clear why it is necessary. So drop it.

[1] https://github.com/systemd/systemd/blob/v258/src/udev/udev-rules.c#L1087

Signed-off-by: Ilya Leoshkevich <iii@linux.ibm.com>