virtme/run: Always mount cache when args.root != '/'
The default CACHE_DIR is $HOME/.cache/virtme-ng. However, that directory is
not available when using an external root filesystem via `--root`. In this
case, it is currently only made available to the guest when used with
`--systemd` but it might be useful in other situations, e.g. with `--ssh`
which also uses the cache for sshd configuration files.

Also, grant write permissions to it, as some things might only be written
within the guest system at runtime, e.g. ssh-keygen.