run: dump kernel console on stdout even if graphics are used
Set up the primary console on hvc0 and connect it to stdio even when
graphics (either `--fb` or `--graphics`) are used.

The hvc0 console is configured as primary (`console=hvc0`), but an
additional cmdline parameter is passed (`virtme_console=tty1`) to
the init to keep the userspace session on the graphical console.

This decouples the ability to inspect the kernel boot log or other
console messages from the fact that the video output is (not) used.
Previously, if a video output of any kind was used, the serial console
on stdio was fully inhibited.

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>