run: only append `debug` to the kernel cmdline if verbosity >= 2
On single `--verbose`, only pass `loglevel=7` rather than `debug`.
On `--verbose --verbose`, pass `debug` as previously.

This makes it easier to view the kernel boot log on the stdio console
(as well as the early console log, which is also tied to `--verbose`)
without also suffering the effects of the debug cmdline option (which
creates much more output and also potentially affects the userspace).

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>