virtme_ng_init: install override into /etc/sudoers.d if supported
If the `/etc/sudoers.d` directory exists and is included from the
already-existing `/etc/sudoers` file, inject virtme-specific sudoers
configuration as a drop-in rather than replacig the primary config.

This way, distro- or user-specific customizations of sudoers config
are preserved (e.g., custom `env_keep` statements which might be
important to the user's workflow).

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>