virtme_ng_init: do not override $PATH if adding a sudoers.d dropin
If we are installing virtme-specific sudoers configuration as a drop-in,
skip injecting a `secure_path`.

Injecting a reasonable typical `secure_path` makes sense if we are
creating a config from scratch, but if we are adding to a pre-existing
config via a drop-in, it is fair to assume that a `secure_path` would be
already set (or unset) according to the user's desires or expectations.

More specifically, the user might have elected to not use `secure_path`
which might be important to their workflow (e.g. by relying on binaries
outside of that path), and we should not override that decision.

Signed-off-by: Ivan Shapovalov <intelfx@intelfx.name>